backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
pandas>=2.2.0
numpy>=1.26.0
sqlalchemy>=2.0.23
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-multipart>=0.0.6
aiosqlite>=0.19.0
reportlab>=4.0.0
python-docx>=1.1.0
matplotlib>=3.8.0
Pillow>=10.0.0
openpyxl>=3.1.0
rapidfuzz>=3.0.0
slowapi>=0.1.9
alembic>=1.13.0

